refactor: restructure and migrate pages to pages dir
