+++ src/App.js
@@ -1,17 +1,28 @@
-import React, { Fragment } from "react";
+import React from "react";
 import "./styles/App.scss";
-import { BrowserRouter } from "react-router-dom";
+import { BrowserRouter, Routes, Route } from "react-router-dom";
+import { Metrics, SMS, NewSMS, Modems, Settings } from "./pages";
+import { AppProvider } from "./store";
 import { Toaster } from "react-hot-toast";
-import { DashBoard } from "./content";
+import { Navbar } from "components";
 
 const App = () => {
   return (
-    <Fragment>
+    <AppProvider>
       <Toaster position="bottom-right" reverseOrder={false} />
       <BrowserRouter>
-        <DashBoard />
+        <Navbar />
+        <div className="bx--col-lg-16 main-content">
+          <Routes>
+            <Route exact path="/" element={<Metrics />} />
+            <Route exact path="/sms" element={<SMS />} />
+            <Route exact path="/new-sms" element={<NewSMS />} />
+            <Route exact path="/modems" element={<Modems />} />
+            <Route exact path="/settings" element={<Settings />} />
+          </Routes>
+        </div>
       </BrowserRouter>
-    </Fragment>
+    </AppProvider>
   )
 }
 export default App;
