--- src/content/Dashboard/index.js
@@ -1,2 +1,0 @@
-import DashBoard from './DashBoard';
-export default DashBoard;
