--- src/content/Metrics/index.js
@@ -1,2 +1,0 @@
-import Metrics from './Metrics';
-export default Metrics;
