--- src/content/Modem/MockData.js
@@ -1,30 +1,0 @@
-
-const MockData = [
-  {
-    id: 1,
-    brand: "Huawei",
-    model: "E303",
-    IMEI: "805485056577889"
-  },
-  {
-    id: 2,
-    brand: "Vodafone",
-    model: "SH-120",
-    IMEI: "805487072577895"
-  },
-  {
-    id: 3,
-    brand: "Huawei",
-    model: "E3131s",
-    IMEI: "805487048577894"
-  },
-  {
-    id: 4,
-    brand: "Vodafone",
-    model: "KE-36",
-    IMEI: "805487056577784"
-  },
-
-];
-
-export default MockData;
