--- src/content/Modem/index.js
@@ -1,2 +1,0 @@
-import Modems from './Modems';
-export default Modems;
