--- src/content/NewSMS/index.js
@@ -1,2 +1,0 @@
-import NewSMS from './NewSMS';
-export default NewSMS;
