--- src/content/SMS/index.js
@@ -1,2 +1,0 @@
-import SMS from './SMS';
-export default SMS;
