--- src/content/Settings/index.js
@@ -1,2 +1,0 @@
-import Settings from './Settings';
-export default Settings;
