+++ src/pages/index.js
@@ -1,0 +1,6 @@
+export { default as DashBoard } from "../components/Navbar";
+export { default as Metrics } from "./Metrics";
+export { default as SMS } from "./SMS";
+export { default as NewSMS } from "./NewSMS";
+export { default as Modems } from "./Modems";
+export { default as Settings } from "./Settings";
