+++ src/styles/App.scss
@@ -18,8 +18,7 @@ $carbon--theme: $carbon--theme--g10;
 
 .main-content {
   height: 100%;
-  margin: $layout-03, 0;
-  margin-top: $layout-04;
+  margin: $layout-04 0 $layout-04 0;
   padding: $spacing-06;
   width: 100%;
 }
